==> 5 the_winter_ball_salad.py <==
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Created on Mon Sep 10 20:11:59 2018

@author: chase.kusterer

Working Directory:
/Users/chase.kusterer/Desktop/PyCourse

Purpose:
Case study to help learn while loops
"""

"""
It's the day before the Winter Ball and our cohort has been tasked with
preparing the garden salads. We have been given a recipe and a need to
make enough for at least 300 people.

Design a loop to help solve this task.
"""


"""
Recipe:
    2 cups of lettuce
    4 pieces of broccoli
    ¼ cup of sliced carrots
    ½ cup of cucumber
    8 cherry tomatoes

Current Inventory:
    50  cups of lettuce
    100 pieces of broccoli
    25  cups of sliced carrots
    25  cups of cucumber
    400 cherry tomatoes


Question 1: How many salads can we make with these ingredients?
Question 2: How much of each ingredient will we have left over?
"""


# assigning the variables
lettuce         = 0
broccoli        = 0
sliced_carrots  = 0
cucumber        = 0
cherry_tomatoes = 0
salads          = 0




# create a while loop
while _____ and ________ and ...... :


print(f"You can make a total of {salads} salads. \n")



# create a conditional statement for each ingredient
if _______:
    print(f"You have a total of {lettuce} cups of lettuce remaining. \n")
else:
    print("You have run out of lettuce. \n")

......




"""
Adjust the ingredient inventories  below so that we have enough to
make 300 salads.

Then run the code above with these new inventory levels.
"""

# New inventories
lettuce         = 0
broccoli        = 0
sliced_carrots  = 0
cucumber        = 0
cherry_tomatoes = 0

# Make sure to set salads back to zero
salads = 0


"""
Prof. Chase:
    What you see above in the new inventory variable assignments is one
    of the beauties of using Python.
"""
